\name{axis.i2wt}
\alias{axis.i2wt}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{specialized axis function for interaction2wt.}
\description{
  Labels the bottom axis with the x-factor name for each column.
  Labels the right axis with the response variable name in all rows.
}
\usage{
axis.i2wt(side, scales, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{side, scales, \dots}{See \code{\link[lattice]{axis.default}}.}
}
\author{Richard M. Heiberger, with asssistance from Deepayan Sarkar.}
\seealso{\code{\link{interaction2wt}}}
\keyword{dplot}
