\name{defunct}
\alias{HH-defunct}
\alias{anova.mean}
\alias{vcov.sufficient}
\alias{aov.sufficient}
\alias{print.glht.mmc.multicomp}
\alias{coef.arima.HH}
\alias{glht.mmc}
\alias{odds.ratio}
\alias{plot.odds.ratio}
\alias{persp.plane}
\alias{persp.plane}
\alias{persp.floor}
\alias{persp.back.wall.x}
\alias{persp.back.wall.y}
\alias{persp.setup}
\alias{plot.hov}
\alias{plot.hov.bf}
\alias{plot.matchMMC}
\alias{seqplot.forecast}
\alias{panel.likert}
\alias{lm.case}

\title{Defunct Functions in Package 'HH'}

\description{ The functions listed here are no longer part of the HH
  package.  Their task has been assigned to different function names.
}

\usage{
anova.mean(...)
vcov.sufficient(...)
aov.sufficient(...)
print.glht.mmc.multicomp(...)
coef.arima.HH(...)
glht.mmc(...)
odds.ratio(...)
plot.odds.ratio(...)
persp.plane(...)
persp.plane(...)
persp.floor(...)
persp.back.wall.x(...)
persp.back.wall.y(...)
persp.setup(...)
plot.hov(...)
plot.hov.bf(...)
plot.matchMMC(...)
seqplot.forecast(...)
panel.likert(...)
lm.case(...)
}
\arguments{
\item{\dots}{other arguments.}
}
\details{Some of these function names have been replaced by using them
as methods.  Some have had their spelling changed to remove the '.'
character.
}
\author{ Richard M. Heiberger <rmh@temple.edu> }
\keyword{misc }
