% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesa_hagstofu.R
\name{lesa_hagstofu}
\alias{lesa_hagstofu}
\title{Lesum inn CSV skrár sem fengnar eru af Hagstofuvefnum}
\usage{
lesa_hagstofu(file)
}
\arguments{
\item{file}{}
}
\value{

}
\description{
Flýtiwrapper utan um read_delim, sérstaklega fyrir CSV skrár af Hagstofuvefnum.
}
\examples{
lesa_hagstofu("MAN03200.csv")
}
