\name{lm.regsubsets}
\alias{lm.regsubsets}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Evaluate lm model with highest adjusted $R^2$.}
\description{
The \code{regsubsets} function in the \code{leaps} package finds the
model with the highest adjusted \eqn{R^2}.  This function evaluates the full
\code{lm} object for that model.
}
\usage{
lm.regsubsets(object, model.number, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{ An object of class \code{"regsubsets"}. }
  \item{model.number}{Index number generated by \code{Rcmdr}.}
  \item{\dots}{Other arguments.}
}
\value{
\code{"lm"} object for the selected model.
}
\author{ Richard M. Heiberger <rmh@temple.edu> }
\seealso{\code{\link[stats]{lm}},  \code{\link[leaps]{regsubsets}}
%%  , \code{\link[RcmdrPlugin.HH]{bestSubsetsRegressionModel.HH}}
}
\keyword{regression}
