\name{strip.background0}
\alias{strip.background0}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Turn off the coloring in the trellis strip labels.  Color 0 is the
 background color.}
\description{
 Turn off the coloring in the trellis strip labels.  Color 0 is the
 background color.
}
\usage{
strip.background0()
}
%- maybe also 'usage' for other objects documented here.
\author{ Richard M. Heiberger <rmh@temple.edu> }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{dplot}
