\name{datasets}
\alias{datasets}
\alias{AEdata}
\alias{AudiencePercent}
\alias{Design_2.8_2}
\alias{Design_2.8_2_full}
\alias{NZScienceTeaching}
\alias{PoorChildren}
\alias{ProfChal}
\alias{ProfDiv}
\alias{SFF8121}
\alias{abc}
\alias{abrasion}
\alias{acacia}
\alias{aeanonym}
\alias{animal}
\alias{anneal}
\alias{apple}
\alias{ara}
\alias{balance}
\alias{barleyp}
\alias{batch}
\alias{bean}
\alias{birthweight}
\alias{blood}
\alias{blyth}
\alias{breast}
\alias{budworm}
\alias{byss}
\alias{c3c4}
\alias{catalystm}
\alias{cc135}
\alias{cc176}
\alias{cement}
\alias{census4}
\alias{cereals}
\alias{chimp}
\alias{circuit}
\alias{co2}
\alias{concord}
\alias{crash}
\alias{crime}
\alias{darwin}
\alias{diamond}
\alias{display}
\alias{distress}
\alias{draft}
\alias{draft70mn}
\alias{drunk}
\alias{eggs}
\alias{elnino}
\alias{employM16}
\alias{energy}
\alias{esr}
\alias{fabricwear}
\alias{fat.dat}
\alias{fat}
\alias{feed}
\alias{filmcoat}
\alias{filter}
\alias{fruitflies}
\alias{furnace}
\alias{girlht}
\alias{glasses}
\alias{golf}
\alias{gum}
\alias{gunload}
\alias{har1}
\alias{har2}
\alias{har3}
\alias{hardness}
\alias{heartvalve}
\alias{hooppine}
\alias{hospital}
\alias{hotdog}
\alias{houseprice}
\alias{hpErie}
\alias{htwt}
\alias{iceskate}
\alias{icu}
\alias{income}
\alias{inconsistent}
\alias{intubate}
\alias{ironpot}
\alias{kangaroo}
\alias{kidney}
\alias{kyphosis}
\alias{lake}
\alias{leukemia}
\alias{lft.asat}
\alias{lifeins}
\alias{longley}
\alias{lymph}
\alias{maiz}
\alias{manhours}
\alias{market}
\alias{mice}
\alias{mileage}
\alias{mortality}
\alias{mpg}
\alias{muscle}
\alias{njgolf}
\alias{normtemp}
\alias{notch}
\alias{nottem}
\alias{oats}
\alias{odoffna}
\alias{operator}
\alias{ozone}
\alias{paper}
\alias{patient}
\alias{plasma}
\alias{political}
\alias{potency}
\alias{pox}
\alias{product}
\alias{psycho}
\alias{pulmonary}
\alias{pulse}
\alias{R282}
\alias{radioact}
\alias{rent}
\alias{retard}
\alias{rhiz.alfalfa}
\alias{rhiz.clover}
\alias{rhizobium1}
\alias{rhizobium3}
\alias{salary}
\alias{salinity}
\alias{salk}
\alias{seeding}
\alias{selfexam}
\alias{shipment}
\alias{sickle}
\alias{skateslc}
\alias{smokers}
\alias{spacshu}
\alias{spindle}
\alias{sprint}
\alias{stopdist}
\alias{surface}
\alias{tablet1}
\alias{teachers}
\alias{testing}
\alias{testscore}
\alias{tires}
\alias{tongue}
\alias{tser.mystery.X}
\alias{tser.mystery.Y}
\alias{tser.mystery.Z}
\alias{tsq}
\alias{turkey}
\alias{tv}
\alias{usair}
\alias{uscrime}
\alias{vocab}
\alias{vulcan}
\alias{washday}
\alias{water}
\alias{weightloss}
\alias{weld}
\alias{wheat}
\alias{wool}
\alias{workstation}
\alias{yates}
\alias{yatesppl}


\title{Datasets for Statistical Analysis and Data Display, Heiberger and
Holland}
\description{
Most of the datasets are described in the book \emph{Statistical Analysis and Data
  Display}.

For \code{ProfChal}, see \code{\link{plot.likert}}.

\code{AudiencePercent} is from personal communication by the market
researcher who did the study.

\code{SFF8121} is student evaluations of my class compared to the
average of all graduate classes in the Spring 2010 semester.
Personal communication from the Temple University
Office of the Provost to me.

\code{ProfDiv} is "Profit-and-Dividend Status of 348 Corportations in
the United States for the period from 1929 to 1935" from
Brinton WC (1939),
 \emph{Graphic Presentation}.
 Brinton Associates.
  \url{http://www.archive.org/details/graphicpresentat00brinrich}.

  \code{NZScienceTeaching} is from
  New Zealand Ministry of Research Science and Technology(2006),
 ``Staying in Science."
\url{http://www.morst.govt.nz/Documents/publications/researchreports/Staying-in-Science-summary.pdf}.

\code{PoorChildren} is from
``Poor Children, Working Parents",
Analysis of data from the CensusBureau's American Community Survey.
Comparison of Census areas of 100,000 or more people,
based on samples from 2005 to 2009.

Source: Data from the U.S. Census Bureau's American Community Survey;
analysis by Andrew A. Beveridge, QueensCollege.
Copyright 2011 The New York Times Company

\url{http://www.nytimes.com/imagepages/2011/12/03/opinion/03blow-ch.html?ref=opinion}

\url{http://www.nytimes.com/2011/12/03/opinion/blow-newts-war-on-poor-children.html?_r=1}

Naomi Robbins and I discuss the PoorChildren example in the Forbes online column:
\url{http://www.forbes.com/sites/naomirobbins/2011/12/20/alternative-to-charles-blows-figure-in-newts-war-on-poor-children-2/}
\code{demo(PoorChildren, package="HH")}
}
\author{ Richard M. Heiberger <rmh@temple.edu> }
\references{
Heiberger, Richard~M. and Holland, Burt (2004b).
\emph{Statistical Analysis and Data Display: An Intermediate Course
  with Examples in S-Plus, R, and SAS}.
 Springer Texts in Statistics. Springer.
 ISBN 0-387-40270-5.
}
\keyword{ datasets }
