\name{panel.isomeans}
\alias{panel.isomeans}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
isomeans grid for MMC plots.
}
\description{
isomeans grid for MMC plots.
}
\usage{
panel.isomeans(ybar)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ybar}{
Vector of means.
}
}
\author{ Richard M. Heiberger <rmh@temple.edu> }

\seealso{
See \code{\link{mmc}} for the references and examples.
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ hplot }
